refactor: Player 패키지 변경으로 인한 컴파일 에러 수정
